+++ daemon/persistence/notification.go
@@ -4,12 +4,12 @@ import "time"
 
 //Notification ...Map to table name notifications
 type Notification struct {
-	ID         uint      `db:"id"`
+	ID         int64      `db:"id"`
 	CreatedAt  time.Time `db:"created_at"`
 	UpdatedAt  time.Time `db:"updated_at"`
 	Actions    string    `db:"actions"`
 	Content    string    `db:"content"`
-	ContentID  uint      `db:"content_id"`
+	ContentID  int64      `db:"content_id"`
 	Link       string    `db:"link"`
 	Originator string    `db:"originator"`
 	TargetUser string    `db:"target_user"`
